# Use separate StringMatcher for TryGetInternedString