--- a/src/Esprima/StringPool.cs
+++ b/src/Esprima/StringPool.cs
@@ -74,6 +74,12 @@ internal struct StringPool
     /// <returns>The stored string instance.</returns>
     public string GetOrCreate(ReadOnlySpan<char> value)
     {
+        if (value.Length > 20)
+        {
+            // too long
+            return value.ToString();
+        }
+
         if (_buckets is null)
         {
             Initialize(4);