Own common keys which allow separation from JS words and limiting to them.  Known strings are based on commonity to ensure we get smaller lookup paths for things that actually are common. Removed some test cases for strings that would bloat decision tree and are less common - ideally we would always find discriminator for optimum matching (until NET 7 where it seems to optimize quite nicely with SequenceEquals).

Slight performance boost.

## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|
| Old |ParseProgram|angular-1.2.5|10.453 ms|0.0914 ms|265.6250|109.3750|4,392 KB|
| **New** |	|	| **10.180 ms (-3%)** | **0.0127 ms** | **265.6250 (0%)** | **109.3750 (0%)** | **4,391 KB (0%)** |
| Old |ParseProgram|backbone-1.1.0|1.416 ms|0.0045 ms|41.0156|17.5781|679 KB|
| **New** |	|	| **1.417 ms (0%)** | **0.0044 ms** | **41.0156 (0%)** | **17.5781 (0%)** | **679 KB (0%)** |
| Old |ParseProgram|jquery-1.9.1|8.362 ms|0.0082 ms|218.7500|109.3750|3,770 KB|
| **New** |	|	| **8.118 ms (-3%)** | **0.0218 ms** | **218.7500 (0%)** | **93.7500 (-14%)** | **3,770 KB (0%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|12.925 ms|0.0752 ms|343.7500|171.8750|5,800 KB|
| **New** |	|	| **12.432 ms (-4%)** | **0.0660 ms** | **343.7500 (0%)** | **171.8750 (0%)** | **5,799 KB (0%)** |
| Old |ParseProgram|mootools-1.4.5|6.661 ms|0.0344 ms|187.5000|85.9375|3,095 KB|
| **New** |	|	| **6.526 ms (-2%)** | **0.0062 ms** | **187.5000 (0%)** | **85.9375 (0%)** | **3,095 KB (0%)** |
| Old |ParseProgram|underscore-1.5.2|1.216 ms|0.0127 ms|33.2031|13.6719|574 KB|
| **New** |	|	| **1.218 ms (0%)** | **0.0015 ms** | **33.2031 (0%)** | **15.6250 (+14%)** | **574 KB (0%)** |
| Old |ParseProgram|yui-3.12.0|6.075 ms|0.0233 ms|171.8750|78.1250|2,900 KB|
| **New** |	|	| **6.037 ms (-1%)** | **0.0226 ms** | **171.8750 (0%)** | **78.1250 (0%)** | **2,853 KB (-2%)** |