@lahma 2022-08-20
Thanks for the review and comments. I need to refine this a bit. In my book it's and error if we need to slice/intern a known operator, matching is logical but after that we can do a lookup with operator type etc.