@adams85 2022-08-20
> In my book it's and error if we need to slice/intern a known operator, matching is logical but after that we can do a lookup with operator type etc.

It's true that operators are represented using enums in the AST, however when collection of tokens are enabled in options, the consumer will have them around in string form as well: https://github.com/lahma/esprima-dotnet/blob/82f4d1b4e7a96f3eea69caf4d35a2d293b3186b9/src/Esprima/JavascriptParser.cs#L269-L271

Interning/pooling punctuators in such a case could make a huge difference considering memory footprint.