@lahma 2022-08-20
@adams85 tweaked `GetTokenRaw` in top commit to account for known punctuators. I created an AI for detecting such cases (throw an exception for known punctuators) and checked that nothing should try to slice one without proper reason (template literal seems to be the only thing doing so).

w.r..t. generator output, I would be very welcome to nerd-snipe you to twek it, but it's hidden output anyway so the eyesore is mostly in diff output in unit tests 😉 

## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|
| Old |ParseProgram|angular-1.2.5|10.337 ms|0.0506 ms|265.6250|109.3750|4,392 KB|
| **New** |	|	| **10.321 ms (0%)** | **0.0306 ms** | **250.0000 (-6%)** | **125.0000 (+14%)** | **4,334 KB (-1%)** |
| Old |ParseProgram|backbone-1.1.0|1.452 ms|0.0072 ms|41.0156|17.5781|679 KB|
| **New** |	|	| **1.391 ms (-4%)** | **0.0080 ms** | **41.0156 (0%)** | **15.6250 (-11%)** | **679 KB (0%)** |
| Old |ParseProgram|jquery-1.9.1|8.446 ms|0.0136 ms|218.7500|109.3750|3,770 KB|
| **New** |	|	| **8.276 ms (-2%)** | **0.0416 ms** | **218.7500 (0%)** | **109.3750 (0%)** | **3,773 KB (0%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|12.632 ms|0.0260 ms|343.7500|171.8750|5,800 KB|
| **New** |	|	| **12.572 ms (0%)** | **0.0844 ms** | **343.7500 (0%)** | **171.8750 (0%)** | **5,817 KB (0%)** |
| Old |ParseProgram|mootools-1.4.5|6.607 ms|0.0087 ms|187.5000|85.9375|3,095 KB|
| **New** |	|	| **6.578 ms (0%)** | **0.0324 ms** | **187.5000 (0%)** | **85.9375 (0%)** | **3,097 KB (0%)** |
| Old |ParseProgram|underscore-1.5.2|1.230 ms|0.0032 ms|33.2031|13.6719|574 KB|
| **New** |	|	| **1.224 ms (0%)** | **0.0032 ms** | **33.2031 (0%)** | **15.6250 (+14%)** | **574 KB (0%)** |
| Old |ParseProgram|yui-3.12.0|6.149 ms|0.0327 ms|171.8750|78.1250|2,900 KB|
| **New** |	|	| **6.057 ms (-1%)** | **0.0205 ms** | **171.8750 (0%)** | **85.9375 (+10%)** | **2,861 KB (-1%)** |