@adams85 2022-08-20
I did some further tweaking, got the following results (compared to the previous commit):

## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|
| Old |ParseProgram|angular-1.2.5|18.644 ms|2.6458 ms|718.7500|343.7500|4,334 KB|
| **New** |	|	| **18.904 ms (+1%)** | **2.6575 ms** | **718.7500 (0%)** | **343.7500 (0%)** | **4,333 KB (0%)** |
| Old |ParseProgram|backbone-1.1.0|2.590 ms|0.1673 ms|140.6250|39.0625|679 KB|
| **New** |	|	| **2.552 ms (-1%)** | **0.2609 ms** | **144.5313 (+3%)** | **35.1563 (-10%)** | **679 KB (0%)** |
| Old |ParseProgram|jquery-1.9.1|15.165 ms|1.8061 ms|609.3750|296.8750|3,773 KB|
| **New** |	|	| **15.108 ms (0%)** | **0.2451 ms** | **625.0000 (+3%)** | **312.5000 (+5%)** | **3,725 KB (-1%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|23.471 ms|0.7369 ms|1000.0000|437.5000|5,817 KB|
| **New** |	|	| **23.349 ms (-1%)** | **0.3698 ms** | **1000.0000 (0%)** | **437.5000 (0%)** | **5,817 KB (0%)** |
| Old |ParseProgram|mootools-1.4.5|11.777 ms|3.7849 ms|500.0000|250.0000|3,097 KB|
| **New** |	|	| **12.028 ms (+2%)** | **2.6814 ms** | **500.0000 (0%)** | **250.0000 (0%)** | **3,097 KB (0%)** |
| Old |ParseProgram|underscore-1.5.2|2.246 ms|0.3332 ms|121.0938|42.9688|574 KB|
| **New** |	|	| **2.240 ms (0%)** | **0.1100 ms** | **117.1875 (-3%)** | **46.8750 (+9%)** | **573 KB (0%)** |
| Old |ParseProgram|yui-3.12.0|11.378 ms|1.2225 ms|484.3750|234.3750|2,861 KB|
| **New** |	|	| **11.352 ms (0%)** | **1.2432 ms** | **468.7500 (-3%)** | **234.3750 (0%)** | **2,860 KB (0%)** |

I hoped for a bit of improvement in execution time but it seems that it doesn't make a difference. (I measured ROS-based `IsKeyword` too but interestingly that performed worse...) Managed to shave off some KBs with [this condition](https://github.com/sebastienros/esprima-dotnet/pull/324/commits/f54429b57aa5892e3c59f8d57aa324afc4fddf85#diff-090f3672b7aac64857879fd490cea99c82fca825dbd87565a7fb7d5bd384a154R37-R40) though.