@adams85 2022-08-21
> I also tweaked the interner to drop matches over 20 chars which would suggest a lot of entropy and probably a bad match (comments did hit this).

Good point, makes perfect sense. However, after reconsidering the problem, I realized that we don't want to limit the length of strings to intern in all cases. I think long identifiers should be interned as identifiers tend to occur more than once. In most of the cases this will save memory. (And think not only of allocations but also of memory footprint of the resulting AST.)

https://github.com/sebastienros/esprima-dotnet/pull/324/commits/7fb617eaa9ca7a37e691e07f954d071b1a52116a implements this.

Benchmark results (compared to your last commit):

## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |ParseProgram|angular-1.2.5|18.644 ms|2.6458 ms|718.7500|343.7500|-|4,334 KB|
| **New** |	|	| **18.796 ms (+1%)** | **1.4861 ms** | **718.7500 (0%)** | **343.7500 (0%)** | **-** | **4,321 KB (0%)** |
| Old |ParseProgram|backbone-1.1.0|2.590 ms|0.1673 ms|140.6250|39.0625|-|679 KB|
| **New** |	|	| **2.552 ms (-1%)** | **0.0741 ms** | **144.5313 (+3%)** | **42.9688 (+10%)** | **-** | **679 KB (0%)** |
| Old |ParseProgram|jquery-1.9.1|15.165 ms|1.8061 ms|609.3750|296.8750|-|3,773 KB|
| **New** |	|	| **14.814 ms (-2%)** | **2.7274 ms** | **609.3750 (0%)** | **296.8750 (0%)** | **-** | **3,769 KB (0%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|23.471 ms|0.7369 ms|1000.0000|437.5000|-|5,817 KB|
| **New** |	|	| **23.617 ms (+1%)** | **0.2659 ms** | **1000.0000 (0%)** | **468.7500 (+7%)** | **31.2500** | **5,804 KB (0%)** |
| Old |ParseProgram|mootools-1.4.5|11.777 ms|3.7849 ms|500.0000|250.0000|-|3,097 KB|
| **New** |	|	| **12.049 ms (+2%)** | **2.7593 ms** | **500.0000 (0%)** | **250.0000 (0%)** | **-** | **3,094 KB (0%)** |
| Old |ParseProgram|underscore-1.5.2|2.246 ms|0.3332 ms|121.0938|42.9688|-|574 KB|
| **New** |	|	| **2.156 ms (-4%)** | **0.0645 ms** | **121.0938 (0%)** | **39.0625 (-9%)** | **-** | **573 KB (0%)** |
| Old |ParseProgram|yui-3.12.0|11.378 ms|1.2225 ms|484.3750|234.3750|-|2,861 KB|
| **New** |	|	| **11.174 ms (-2%)** | **1.0143 ms** | **468.7500 (-3%)** | **234.3750 (0%)** | **-** | **2,860 KB (0%)** |