@lahma 2022-08-21
Thanks for your help on this. I ran one final benchmark between main and this branch, here's the diff tool output:

## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|
| Old |ParseProgram|angular-1.2.5|10.045 ms|0.0221 ms|265.6250|109.3750|4,392 KB|
| **New** |	|	| **9.870 ms (-2%)** | **0.0263 ms** | **250.0000 (-6%)** | **109.3750 (0%)** | **4,320 KB (-2%)** |
| Old |ParseProgram|backbone-1.1.0|1.412 ms|0.0054 ms|41.0156|17.5781|679 KB|
| **New** |	|	| **1.337 ms (-5%)** | **0.0032 ms** | **41.0156 (0%)** | **15.6250 (-11%)** | **679 KB (0%)** |
| Old |ParseProgram|jquery-1.9.1|8.230 ms|0.0205 ms|218.7500|109.3750|3,770 KB|
| **New** |	|	| **7.762 ms (-6%)** | **0.0219 ms** | **218.7500 (0%)** | **78.1250 (-29%)** | **3,769 KB (0%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|12.371 ms|0.0418 ms|343.7500|171.8750|5,800 KB|
| **New** |	|	| **12.160 ms (-2%)** | **0.0249 ms** | **343.7500 (0%)** | **171.8750 (0%)** | **5,803 KB (0%)** |
| Old |ParseProgram|mootools-1.4.5|6.445 ms|0.0118 ms|187.5000|85.9375|3,095 KB|
| **New** |	|	| **6.299 ms (-2%)** | **0.0159 ms** | **187.5000 (0%)** | **85.9375 (0%)** | **3,094 KB (0%)** |
| Old |ParseProgram|underscore-1.5.2|1.208 ms|0.0076 ms|33.2031|13.6719|574 KB|
| **New** |	|	| **1.131 ms (-6%)** | **0.0028 ms** | **33.2031 (0%)** | **15.6250 (+14%)** | **573 KB (0%)** |
| Old |ParseProgram|yui-3.12.0|6.007 ms|0.0155 ms|171.8750|78.1250|2,900 KB|
| **New** |	|	| **5.742 ms (-4%)** | **0.0091 ms** | **171.8750 (0%)** | **85.9375 (+10%)** | **2,860 KB (-1%)** |

I'm happy with this, should we merge?